Modularising tests
Splitting up monolithic test suite into corresponding modules.
