+++ test/common.js
@@ -1,0 +1,9 @@
+var Buffer = require("buffer/").Buffer;
+var crypto_lib = require("crypto-browserify");
+var should = require("should");
+
+var lisk = require("../index.js");
+
+exports.lisk = lisk;
+exports.crypto_lib = crypto_lib;
+exports.should = should;
