+++ test/transactions/multisignature.js
@@ -1,0 +1,7 @@
+var common = require("../common");
+var lisk = common.lisk;
+var should = common.should;
+var crypto_lib = common.crypto_lib;
+
+describe("multisignature.js", function () {
+});
